refactors diagnostic reports controller
