+++ app/queries/recommendations.rb
@@ -1,0 +1,133 @@
+class Recommendations
+    def diagnostic
+        [
+            {
+                recommendation: 'Fragments',
+                activityPackId: 3,
+                requirements: [
+                    {
+                        concept_id: 'j89kdRGDVjG8j37A12p37Q',
+                        count: 3
+                    },
+                    {
+                        concept_id: 'KfA8-dg8FvlJz4eY0PkekA',
+                        count: 6
+                    },
+                    {
+                        concept_id: 'LH3szu784pXA5k2N9lxgdA',
+                        count: 3
+                    }
+                ]
+            },
+
+            {
+                recommendation: 'Compound Subjects, Objects, and Predicates',
+                activityPackId: 6,
+                requirements: [
+                    {
+                        concept_id: 'Jl4ByYtUfo4VhIKpMt23yA',
+                        count: 1
+                    },
+                    {
+                        concept_id: 'QNkNRs8zbCXU7nLBeo4mgA',
+                        count: 1
+                    }
+                ]
+            },
+
+            {
+                recommendation: 'Adjectives',
+                activityPackId: 8,
+                requirements: [
+                    {
+                        concept_id: 'oCQCO1_eVXQ2zqw_7QOuBw',
+                        count: 1
+                    }
+                ]
+            },
+
+            {
+                recommendation: 'Adverbs',
+                activityPackId: 7,
+                requirements: [
+                    {
+                        concept_id: 'GZ04vHSTxWUTzhWMGfwcUQ',
+                        count: 1
+                    }
+                ]
+            },
+
+            {
+                recommendation: 'Compound Sentences',
+                activityPackId: 4,
+                requirements: [
+                    {
+                        concept_id: 'GiUZ6KPkH958AT8S413nJg',
+                        count: 2
+                    },
+                    {
+                        concept_id: 'Qqn6Td-zR6NIAX43NOHoCg',
+                        count: 1
+                    },
+                    {
+                        concept_id: 'hJKqVOkQQQgfEsmzOWC1xw',
+                        count: 1
+                    },
+                    {
+                        concept_id: 'kODm7xLmhGShct5uJIfHXg',
+                        count: 1
+                    }
+                ]
+            },
+
+            {
+                recommendation: 'Complex Sentences',
+                activityPackId: 5,
+                requirements: [
+                    {
+                        concept_id: 'nb0JW1r5pRB5ouwAzTgMbQ',
+                        count: 2
+                    },
+                    {
+                        concept_id: 'Q8FfGSv4Z9L2r1CYOfvO9A',
+                        count: 2
+                    },
+                    {
+                        concept_id: 'S8b-N3ZrB50CWgxD5yg8yQ',
+                        count: 1
+                    },
+                    {
+                        concept_id: '7H2IMZvq0VJ4Uvftyrw7Eg',
+                        count: 1
+                    }
+                ]
+            },
+
+            {
+                recommendation: 'Appositive and Modifying Phrases',
+                activityPackId: 2,
+                requirements: [
+                    {
+                        concept_id: 'InfGdB6Plr2M930kqsn63g',
+                        count: 1
+                    },
+                    {
+                        concept_id: 'GLjAExmqZShBTZ7DQGvVLw',
+                        count: 1
+                    }
+                ]
+            },
+
+            {
+                recommendation: 'Parallel Structure',
+                activityPackId: 10,
+                requirements: [
+                    {
+                        concept_id: '1ohLyApTz7lZ3JszrA98Xg',
+                        count: 1
+                    }
+                ]
+            }
+        ]
+    end
+end
